requirements.txt
cycler==0.10.0
kiwisolver==1.3.1
matplotlib==3.3.3
numpy==1.19.4
opencv-python==4.4.0.46
Pillow==8.0.1
pyparsing==2.4.7
pytesseract==0.3.6
python-dateutil==2.8.1
six==1.15.0
